+++ docs/releases.md
@@ -1,5 +1,15 @@
 # Releases for Audience Manager API python wrapper
 
+## Version 0.0.3
+
+* Adding new methods for Algorithmic traits:
+  * getModels
+  * getModel
+  * deleteModel
+  * getModelTraits
+  * getModelStats
+* clean the createConfigFile method.
+
 ## Version 0.0.2
 
 * Fix issue on delete Trait method.